"Finished user profile page for Milestone 1, everything setup to make a pull request, testing still to be done"
